% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusion_exclusion_criteria.R
\name{alternate_criteria}
\alias{alternate_criteria}
\title{For altering inclusion criteria, to see how changing the criteria affects number of patients}
\usage{
alternate_criteria(df, x, criteria)
}
\arguments{
\item{df}{a dataframe, or tibble}

\item{x}{variable that is used for inclusion/exclusion criteria}

\item{criteria}{value of criteria being applied (can be inclusion or exclusion)}
}
\value{

}
\description{
For altering inclusion criteria, to see how changing the criteria affects number of patients
}
