% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcare_analyses.R
\name{annualized}
\alias{annualized}
\title{Annualize variable}
\usage{
annualized(df, x, var_months)
}
\arguments{
\item{df}{A dataframe or tibble}

\item{x}{variable to be converted to an annualized estimate}

\item{var_months}{variable used indicating the number of months}
}
\value{
variable annualized (per variable per year)
}
\description{
Annualize variable
}
