% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cleaning.R
\name{cleaneR_string}
\alias{cleaneR_string}
\title{Clean a String, trim whitespace and all lowercase}
\usage{
cleaneR_string(x)
}
\arguments{
\item{x}{a vector}
}
\value{
vector x, in all lowercase and with whitespace trimmed
}
\description{
Clean a String, trim whitespace and all lowercase
}
