% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{count_percent}
\alias{count_percent}
\title{Calculate count and proportion for variables}
\usage{
count_percent(df, x)
}
\arguments{
\item{df}{a dataframe or tibble}

\item{x}{variable to get count and percent of}
}
\value{
a dataframe with number and percent of each group
}
\description{
Calculate count and proportion for variables
}
