% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{count_percent_demo}
\alias{count_percent_demo}
\title{Calculate n and \% for unique pt ids}
\usage{
count_percent_demo(df, x, pt_id)
}
\arguments{
\item{df}{dataframe}

\item{x}{variable you want to group by}

\item{pt_id}{participant id}
}
\description{
Calculate n and \% for unique pt ids
}
