% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{mean_by_group}
\alias{mean_by_group}
\title{Mean by group}
\usage{
mean_by_group(df, x, grouping_var)
}
\arguments{
\item{df}{a dataframe}

\item{x}{a vector / variable to calculate the mean and SD for}

\item{grouping_var}{a vector/variable to group by}
}
\value{
the mean and sd for each group of the specified variable
}
\description{
Mean by group
}
