% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculate Mean and Standard Deviation}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{a numeric vector x (ie, df$x)}
}
\value{
returns the mean (SD) rounded to 2 decimal places
}
\description{
Calculate Mean and Standard Deviation
}
