% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{median_by_group}
\alias{median_by_group}
\title{Median by group}
\usage{
median_by_group(df, x, grouping_var)
}
\arguments{
\item{df}{a dataframe}

\item{x}{a vector / variable to calculate the median and IQR for}

\item{grouping_var}{a vector/variable to group by}
}
\value{
the median and IQR for each group of the specified variable
}
\description{
Median by group
}
