% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{median_iqr}
\alias{median_iqr}
\title{Calculate Median and Interquartile Range (IQR)}
\usage{
median_iqr(x)
}
\arguments{
\item{x}{a numeric vector x (ie, df$x)}
}
\value{
returns the mean (SD) rounded to 2 decimal places
}
\description{
Calculate Median and Interquartile Range (IQR)
}
