% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_counts.R
\name{num_pts_trt}
\alias{num_pts_trt}
\title{Number of Patients with a Treatment}
\usage{
num_pts_trt(df, x, value, id)
}
\arguments{
\item{df}{a dataframe}

\item{x}{a variable}

\item{value}{the treatment being specified, for example: "allergy medication"}

\item{id}{a patient identifier}
}
\value{

}
\description{
Number of Patients with a Treatment
}
