% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_counts.R
\name{num_trt}
\alias{num_trt}
\title{Number of Treatments}
\usage{
num_trt(df, x, value)
}
\arguments{
\item{df}{a dataframe}

\item{x}{a variable}

\item{value}{treatment to look for, for example: "allergy medication"}
}
\value{

}
\description{
Number of Treatments
}
