% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{num_unique}
\alias{num_unique}
\title{Number of Unique Values}
\usage{
num_unique(df, x)
}
\arguments{
\item{df}{dataframe, or tibble.}

\item{x}{vector / variable that you want unique number of (i.e, patient identifier, age group, sex)}
}
\value{
number of unique values
}
\description{
Number of Unique Values
}
