% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{num_unique_pt}
\alias{num_unique_pt}
\title{Counts number of Unique Pts}
\usage{
num_unique_pt(df, id)
}
\arguments{
\item{df}{a dataframe or tibble}

\item{id}{patient id}
}
\description{
Counts number of Unique Pts
}
