% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample_count}
\alias{resample_count}
\title{Counts and Proportions of Resampled Data}
\usage{
resample_count(df, sample_size_drawn)
}
\arguments{
\item{df}{a dataframe containing resampled data}

\item{sample_size_drawn}{size of the sample that was drawn}
}
\value{
returns dataframe with counts and proportions for each sample drawn
}
\description{
Counts and Proportions of Resampled Data
}
