% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample_data}
\alias{resample_data}
\title{Resample Data}
\usage{
resample_data(
  times_to_resample,
  df.vector,
  sample_size_drawn,
  replace_response = TRUE
)
}
\arguments{
\item{times_to_resample}{number of times to resample}

\item{df.vector}{vector to draw samples from}

\item{sample_size_drawn}{number of samples to draw}

\item{replace_response}{sample with replacement. By default, TRUE}
}
\value{
returns a dataframe, resampled n times
}
\description{
Resample Data
}
