% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{unique_values}
\alias{unique_values}
\title{Unique Values}
\usage{
unique_values(df, x)
}
\arguments{
\item{df}{dataframe, or tibble}

\item{x}{vector / variable that you want unique number of (i.e, patient identifier, age group, sex)}
}
\value{
dataframe of unique values (i.e, patient identifiers)
}
\description{
Unique Values
}
