% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{count_percent}
\alias{count_percent}
\title{Calculate n and \% of each group in a variable}
\usage{
count_percent(df, x)
}
\arguments{
\item{df}{a dataframe}

\item{x}{variable you want to group by}
}
\description{
Calculate n and \% of each group in a variable
}
