% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{count_percent_demo}
\alias{count_percent_demo}
\title{Calculate count and percent for unique participant ids}
\usage{
count_percent_demo(df, x, pt_id)
}
\arguments{
\item{df}{a dataframe or tibble}

\item{x}{variable x}

\item{pt_id}{patient identifier}
}
\value{
a dataframe with number and percent of each group
}
\description{
Calculate count and percent for unique participant ids
}
