% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{num_unique_pt}
\alias{num_unique_pt}
\title{Number of Unique Patients}
\usage{
num_unique_pt(df, id)
}
\arguments{
\item{df}{dataframe, or tibble.}

\item{id}{patient identifier}
}
\value{
number of unique patient identifiers
}
\description{
Number of Unique Patients
}
