% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{unique_pt}
\alias{unique_pt}
\title{Unique Patients IDs}
\usage{
unique_pt(df, id)
}
\arguments{
\item{df}{dataframe, or tibble}

\item{id}{patient identifier}
}
\value{
dataframe of unique patient identifiers
}
\description{
Unique Patients IDs
}
